+++ src/main/java/com/bitbank/services/BankAccountTransactionsService.java
@@ -2,6 +2,8 @@ package com.bitbank.services;
 
 import java.math.BigDecimal;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -10,10 +12,16 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import com.bitbank.entities.BankAccountTransactionsEntity;
+import com.bitbank.exceptions.BalanceNotEnoughException;
 import com.bitbank.repositories.BankAccountTransactionsRepository;
 
 @Service
 public class BankAccountTransactionsService {
+
+	/**
+	 * Create our logger
+	 */
+	Logger logger = LoggerFactory.getLogger(BankAccountTransactionsService.class);
 
 	@Autowired
 	private BankAccountTransactionsRepository bankAccountTransactionsRepository;
@@ -38,9 +46,29 @@ public class BankAccountTransactionsService {
 	}
 
 	/**
-	 * Post
+	 * Add or deduct balance.
+	 * 
+	 * @throws BalanceNotEnoughException
 	 */
-	public BankAccountTransactionsEntity post(BankAccountTransactionsEntity bankAccountTransactionsEntity) {
+	public BankAccountTransactionsEntity post(BankAccountTransactionsEntity bankAccountTransactionsEntity)
+			throws BalanceNotEnoughException {
+
+		// Get current balance
+		BigDecimal balance = this.balance();
+		// Get amount
+		BigDecimal amount = bankAccountTransactionsEntity.getAmount();
+		// Arithmetical sum between balance and amount
+		BigDecimal arithmeticalSum = balance.add(amount);
+
+		// Compare data
+		if (arithmeticalSum.compareTo(BigDecimal.ZERO) < 0) {
+			/**
+			 * "Preconditions" and logging arguments should not require evaluation
+			 */
+			logger.info("Cannot deduct {} from {}", amount, balance);
+			throw new BalanceNotEnoughException("balance is not enough to deduct " + amount.toString());
+		}
+
 		return bankAccountTransactionsRepository.save(bankAccountTransactionsEntity);
 	}