+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-# 0.12.0
+# 0.13.0
++ Add `BalanceNotEnoughException`
++ Refactor `BalanceNotEnoughException`
++ Refactor CORS
++ Validate registration
++ Move CORS to application properties
++ Add logger configuration
+
+## 0.12.0
 + Generate token from token
 + Add Refreh Token route
 + Get expiry date from token